requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-groq
langchain-community
langgraph
ddgs
python-dotenv
pydantic
streamlit
